% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv-to-qx.R
\name{calc_qx_from_surv}
\alias{calc_qx_from_surv}
\title{Calculate qx column from a \code{survfit} object}
\usage{
calc_qx_from_surv(obj_survfit, ...)
}
\arguments{
\item{obj_survfit}{An object returned from a call to \code{survival::survfit()} or
\code{survival::survfit0()}.}
}
\value{
A \code{tibble} with five columns - \code{stratum} first, which reports the stratum
from the \code{survfit} object to which the estimate belongs. The second, third,
and fourth columns return standard life table estimates, including
person-years (\code{py}), mortality rates (\code{mx}), and the probability of death
(\code{qx}).
}
\description{
This function takes an object resulting from a call to \code{survival::survfit()}
or \code{survival::survfit0()} object and calculates a single-year lifetable for
each stratum.
}
