% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv-to-qx.R
\name{convert_survfit2tibble}
\alias{convert_survfit2tibble}
\title{Convert estimates from a \code{survfit} object to a tibble (dataframe)}
\usage{
convert_survfit2tibble(
  obj_survfit,
  time_in = c("days", "years"),
  age_max_years = 120L
)
}
\arguments{
\item{obj_survfit}{A \code{survfit} object generated using \code{survival::survfit()}
or \code{survival::survfit0()}.}

\item{time_in}{Unit in which time is reported. Either \code{days} or \code{years}.}

\item{age_max_years}{Maximum age, in years, to which data should be
extrapolated. Defaults to 120.}
}
\value{
A \code{tibble} with \code{age_max_years + 1} rows and three columns. The first column,
\code{year}, gives the year, the second, \code{n_events} the number events (deaths) in
this year, and the third, \code{n_start}, the number at risk at the start of the
year.
}
\description{
This function takes a \code{survfit} object and extracts time estimates, together
with the number of events and number of risk. Based on the user specification
of time units - either \code{days} or \code{years} - the function then creates a full
sequence of time and event/risk number estimates from 0 to a user-specified
maximum age.
}
\details{
Between time points at which an event takes place, the number at
risk is retained from the preceding time point with an event, while the
number of events is zero. Beyond the last observed time, the number at risk
from the last observed time is retained, with no more events taking place.
}
