% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-covalence.R
\name{covalence_colors}
\alias{covalence_colors}
\title{Covalence colors}
\usage{
covalence_colors(...)
}
\arguments{
\item{...}{Access specific colors}
}
\value{
A named vector of brand colors
}
\description{
The Covalence brand colors, including the primary colors
('lightblue', 'darkblue', 'teal', and 'grey') and the accent colors
('red', 'green', 'purple', and 'orange').
}
\examples{
covalence_colors("darkblue")
}
