% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-covalence.R
\name{covalence_colours}
\alias{covalence_colours}
\title{Covalence colours}
\usage{
covalence_colours(...)
}
\arguments{
\item{...}{Access specific colours}
}
\value{
A named vector of brand colours
}
\description{
The Covalence brand colours, including the primary colours
('lightblue', 'darkblue', 'teal', and 'grey') and the accent colours
('red', 'green', 'purple', and 'orange'). Each colour name is prefixed with
"cov_" to avoid confusion with the default R colour names.
}
\details{
This function is taken from a 2022 \href{https://meghan.rbind.io/blog/2022-10-11-creating-custom-color-palettes-with-ggplot2/#defining-custom-colors-and-palettes}{blog post} by Meghan Hall.
}
\examples{
covalence_colours("darkblue")
}
