% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-covalence.R
\name{covalence_palette}
\alias{covalence_palette}
\title{Covalence palette}
\usage{
covalence_palette(..., palette = c("complete", "main", "accent"))
}
\arguments{
\item{...}{Arguments to be passed on.}

\item{palette}{Palette to use. One of 'complete', 'main', or 'accent'.}
}
\value{
A named vector of brand colors.
}
\description{
The various Covalence palettes:
\itemize{
\item The \emph{complete} palette that contains all brand colors.
\item The \emph{main} palette that contains the main, mostly blue-ish, brand colors.
\item The \emph{accent} palette that contains additional, more prominent colors.
}
}
\details{
This function is slightly modified from a 2022 \href{https://meghan.rbind.io/blog/2022-10-11-creating-custom-color-palettes-with-ggplot2/#defining-custom-colors-and-palettes}{blog post} by Meghan Hall.
}
\examples{
scales::show_col(covalence_palette(palette = "main"))
}
