% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-covalence.R
\name{generate_pal}
\alias{generate_pal}
\title{Function factory to generate palette}
\usage{
generate_pal(palette = "main", type = "discrete", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette to use. One of 'complete', 'main', or 'accent'.}

\item{type}{Type of palette. Either 'discrete' or 'continuous'.}

\item{reverse}{Should the palette be reversed? Either 'TRUE' or 'FALSE'.}

\item{...}{Dots.}
}
\value{
A palette-generating function.
}
\description{
This function generates a second function, for use in the 'ggplot' scales.
}
\details{
This function is slightly modified from a 2022 \href{https://meghan.rbind.io/blog/2022-10-11-creating-custom-color-palettes-with-ggplot2/#defining-custom-colors-and-palettes}{blog post} by Meghan Hall.
}
