% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv-to-qx.R
\name{make_lt}
\alias{make_lt}
\title{Make a single-year life table from numbers of events and at risk}
\usage{
make_lt(data)
}
\arguments{
\item{data}{A tibble produced by \code{convert_survfit2tibble()}.}
}
\value{
A \code{tibble} with four columns, the first being \code{year}. The second column,
\code{py}, gives the person-years spent in year \code{x}.The third column, \code{mx}, gives
the mortality rate in year \code{x}; the last, \code{qx}, the probability of dying
between years \code{x} and \code{x+1}.
}
\description{
description
Based on the output from \code{convert_survfit2tibble()}, make a tibble containing
\code{qx} (probabilities of dying before reaching age \code{x+1}) by year.
}
\details{
The person-years (\code{py}) lived in each year are calculated as the mean of the
cohort at age \code{x} and age \code{x+1}, i.e., deaths are assumed to occur,
on average, mid-way through the single-year interval.
The mortality rate, \code{mx}, is calculated as the number of
events (deaths) per person-year.
The probability of death between ages \code{x} and \code{x+1} is calculated as
\eqn{q_{x} = 1 - \exp{-n*m_{x}}}, where \code{n=1} as intervals are single-year.
This implies the assumption of constant mortality in the 1-year interval.
}
