% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-rob2.R
\name{plot_rob2_summary}
\alias{plot_rob2_summary}
\title{Plot summary of RoB2 assessments}
\usage{
plot_rob2_summary(data, domain, judgement, study_share)
}
\arguments{
\item{data}{A data.frame.}

\item{domain}{Unquoted column name for the bias domains.}

\item{judgement}{Unquoted column name for the risk of bias judgement.}

\item{study_share}{Unquoted column name for proportion [0, 1] of studies with
a specific bias risk per domain. Values in this column must all be numeric.}
}
\value{
A ggplot2 object
}
\description{
The \href{https://www.riskofbias.info/welcome/rob-2-0-tool}{RoB2 tool},
developed by the Cochrane Collaboration, allows assessing the
risk of bias in randomized trials. In addition to a study-level traffic light
plot, a typical output of RoB2 assessments is a summary of the proportion
of studies with a specific bias risk, per domain. This latter type of plot is
created by \code{plot_rob2_summary}, which requires summarized data (i.e., no
data synthesis or calculations are performed by the function).
}
\examples{
library(ggplot2)

data_rob2 <- data.frame(
  domain = rep(
    c(
      "Overall bias",
      "Bias due to deviations from intended interventions",
      "Missing data"
    ),
    each = 3
  ),
  judgement = rep(c("Low", "Some concerns", "High risk of bias"), times = 3),
  study_share = c(0.4, 0.4, 0.2, 0.1, 0.85, 0.05, 0.2, 0.5, 0.3)
)

plot_rob2_summary(
  data_rob2,
  domain = domain,
  judgement = judgement,
  study_share = study_share
)
}
