% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-covalence.R
\name{scale_colour_covalence_d}
\alias{scale_colour_covalence_d}
\alias{scale_color_covalence_d}
\alias{scale_fill_covalence_d}
\alias{scale_colour_covalence_c}
\alias{scale_color_covalence_c}
\alias{scale_fill_covalence_c}
\title{Covalence color and fill scale for ggplot2}
\usage{
scale_colour_covalence_d(palette = "complete", reverse = FALSE)

scale_color_covalence_d(palette = "complete", reverse = FALSE)

scale_fill_covalence_d(palette = "complete", reverse = FALSE)

scale_colour_covalence_c(palette = "complete", reverse = FALSE)

scale_color_covalence_c(palette = "complete", reverse = FALSE)

scale_fill_covalence_c(palette = "complete", reverse = FALSE)
}
\arguments{
\item{palette}{Palette to use. One of 'complete', 'main', or 'accent'.}

\item{reverse}{Should the palette be reversed? Either 'TRUE' or 'FALSE'.}
}
\value{
Discrete/continuous color/fill scales for ggplot2.
}
\description{
The \verb{scale_*_covalence_*} functions provide discrete (\verb{_d}) and continuous
(\verb{_c}) scales of Covalence colors for use in \link{ggplot2} plots.
}
\details{
This function is modified from a 2022 \href{https://meghan.rbind.io/blog/2022-10-11-creating-custom-color-palettes-with-ggplot2/#defining-custom-colors-and-palettes}{blog post} by Meghan Hall.
}
\examples{
library(ggplot2)

ggplot(data = diamonds, aes(x = carat, y = price)) +
  geom_point(aes(color = cut), alpha = 0.4) +
  scale_color_covalence_d(palette = "complete") +
  theme_covalence()

ggplot(data = diamonds, aes(x = carat, y = price)) +
  geom_point(aes(color = x), alpha = 0.4) +
  scale_colour_covalence_c(palette = "complete") +
  theme_covalence()

ggplot(data = diamonds, aes(x = price)) +
  geom_histogram(aes(fill = cut)) +
  scale_fill_covalence_d(palette = "complete", reverse = TRUE) +
  theme_covalence()

}
