% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-covalence.R
\name{scale_colour_covalence_discrete}
\alias{scale_colour_covalence_discrete}
\alias{scale_color_covalence_discrete}
\alias{scale_fill_covalence_discrete}
\alias{scale_colour_covalence_c}
\alias{scale_color_covalence_c}
\alias{scale_fill_covalence_c}
\title{Discrete colour and fill scales for ggplot2}
\usage{
scale_colour_covalence_discrete(palette = "complete", reverse = FALSE)

scale_color_covalence_discrete(palette = "complete", reverse = FALSE)

scale_fill_covalence_discrete(palette = "complete", reverse = FALSE)

scale_colour_covalence_c(palette = "complete", reverse = FALSE)

scale_color_covalence_c(palette = "complete", reverse = FALSE)

scale_fill_covalence_c(palette = "complete", reverse = FALSE)
}
\arguments{
\item{palette}{Palette to use. One of 'complete', 'main', 'accent', or 'rainbow_12_bit'.}

\item{reverse}{Should the palette be reversed? Either 'TRUE' or 'FALSE'.}
}
\value{
Discrete colour/fill scales for ggplot2.
}
\description{
The \verb{scale_*_covalence_discrete} functions provide discrete scales of
Covalence colours for use in \link{ggplot2} plots.
}
\details{
These functions are inspired by a 2022 \href{https://www.jumpingrivers.com/blog/custom-colour-palettes-for-ggplot2/}{blog post} by Nicola Rennie.
}
\examples{
library(ggplot2)

ggplot(data = diamonds, aes(x = carat, y = price)) +
  geom_point(aes(colour = cut), alpha = 0.4) +
  scale_colour_covalence_discrete(palette = "complete") +
  theme_covalence()

ggplot(data = diamonds, aes(x = carat, y = price)) +
  geom_point(aes(colour = x), alpha = 0.4) +
  scale_colour_covalence_c(palette = "complete") +
  theme_covalence()

ggplot(data = diamonds, aes(x = price)) +
  geom_histogram(aes(fill = cut)) +
  scale_fill_covalence_discrete(
    palette = "rainbow_12_bit",
    reverse = TRUE
  ) +
  theme_covalence()
}
