% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-covalence.R
\name{scale_colour_covalence_diverging}
\alias{scale_colour_covalence_diverging}
\alias{scale_color_covalence_diverging}
\alias{scale_fill_covalence_diverging}
\title{Covalence diverging colour and fill scales for ggplot2}
\usage{
scale_colour_covalence_diverging(
  palette = "orange_teal",
  reverse = FALSE,
  midpoint = 0,
  na_colour = "#cccccc"
)

scale_color_covalence_diverging(
  palette = "orange_teal",
  reverse = FALSE,
  midpoint = 0,
  na_colour = "#cccccc"
)

scale_fill_covalence_diverging(
  palette = "orange_teal",
  reverse = FALSE,
  midpoint = 0,
  na_colour = "#cccccc"
)
}
\arguments{
\item{palette}{One of the diverging colour palettes (\code{orange_teal},
\code{orange_purple}, or \code{orange_darkblue}). Default is \code{orange_teal}).}

\item{reverse}{Should the palette be reversed? Either 'TRUE' or 'FALSE'.}

\item{midpoint}{Numeric value for midpoint. Default is 0.}

\item{na_colour}{Colour for missing values. Default is "#cccccc" (grey80).}
}
\value{
Diverging colour/fill scales for ggplot2.
}
\description{
The \verb{scale_*_covalence_diverging} functions provide diverging
scales of Covalence colours for use in \link{ggplot2} plots.
}
\details{
These functions are inspired by a 2022 \href{https://rfortherestofus.com/2022/02/data-viz-org-branding/}{blog post} by Cara Thompson.
}
\examples{
library(ggplot2)

}
