% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-covalence.R
\name{scale_colour_covalence_sequential}
\alias{scale_colour_covalence_sequential}
\alias{scale_color_covalence_sequential}
\alias{scale_fill_covalence_sequential}
\title{Sequential colour and fill scales for ggplot2}
\usage{
scale_colour_covalence_sequential(
  palette = "darkblues",
  reverse = FALSE,
  na_colour = "#cccccc"
)

scale_color_covalence_sequential(
  palette = "darkblues",
  reverse = FALSE,
  na_colour = "#cccccc"
)

scale_fill_covalence_sequential(
  palette = "darkblues",
  reverse = FALSE,
  na_colour = "#cccccc"
)
}
\arguments{
\item{palette}{One of the sequential colour palettes (\code{darkblues},
\code{teals}, or \code{purples}). Default is \code{darkblues}).}

\item{reverse}{Should the palette be reversed? Either 'TRUE' or 'FALSE'.}

\item{na_colour}{Colour for missing values. Default is "#cccccc" (grey80).}
}
\value{
Sequential colour/fill scales for ggplot2.
}
\description{
The \verb{scale_*_covalence_sequential} functions provide sequential
scales of Covalence colours for use in \link{ggplot2} plots.
}
\examples{
library(ggplot2)

ggplot(data = diamonds, aes(x = cut, y = carat)) +
  geom_point(aes(colour = price)) +
  scale_colour_covalence_sequential() +
  theme_covalence()

ggplot(data = diamonds, aes(x = cut, y = color)) +
  geom_tile(aes(fill = depth)) +
  scale_fill_covalence_sequential(palette = "teals") +
  theme_covalence() +
  theme(panel.grid.major = element_blank())
}
