% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-covalence.R
\name{theme_covalence}
\alias{theme_covalence}
\title{The default Covalence \link{ggplot2} theme}
\usage{
theme_covalence(
  base_size = 12,
  base_family = "sans",
  strip_bg_colour = "#FFFFFF",
  strip_text_colour = "#326aa0",
  ...
)
}
\arguments{
\item{base_size}{Base font size, in points. Default is 12.}

\item{base_family}{Base font family. Default is "sans".}

\item{strip_bg_colour}{colour to fill strip backgrounds. Default is #326aa0.}

\item{strip_text_colour}{colour for strip labels. Default is #FFFFFF (white).}

\item{...}{Additional arguments to be passed to the theme}
}
\value{
A ggplot2 \link[ggplot2:theme]{theme} object.
}
\description{
This theme provides the defaults required to generate plots, using \link{ggplot2},
that are aligned with the Covalence Research house style. The theme can be
modified further by using \code{theme()}. Also consider using the various colour
palettes to ensure the house style is met.
}
\examples{
library(ggplot2)

ggplot(data = msleep, aes(x = brainwt, y = bodywt)) +
  geom_point() +
  theme_covalence()

ggplot(data = diamonds, aes(x = price, y = carat)) +
  geom_point(colour = covalence_colours("cov_lightblue"), alpha = 0.4) +
  facet_wrap(vars(cut)) +
  theme_covalence()
}
