% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-covalence.R
\name{covalence_palette}
\alias{covalence_palette}
\title{Covalence palette}
\usage{
covalence_palette(
  ...,
  palette = c("discrete_complete", "discrete_main", "discrete_accent",
    "discrete_rainbow_12_bit", "diverging_OrTe")
)
}
\arguments{
\item{...}{Arguments to be passed on.}

\item{palette}{Palette to use. One of 'complete', 'main', 'accent', or 'rainbow_12_bit'.}
}
\value{
A named vector of brand colors.
}
\description{
The various Covalence palettes:
\itemize{
\item The \emph{complete} palette that contains all brand colors.
\item The \emph{main} palette that contains the main, mostly blue-ish, brand colors.
\item The \emph{accent} palette that contains additional, more prominent colors.
}
}
\details{
This function is slightly modified from a 2022 \href{https://meghan.rbind.io/blog/2022-10-11-creating-custom-color-palettes-with-ggplot2/#defining-custom-colors-and-palettes}{blog post} by Meghan Hall.
The \href{https://iamkate.com/data/12-bit-rainbow/}{12-bit rainbow palette} was developed by Kate Morley and released under the terms of the Creative Commons CC0 1.0 Universal Legal Code.
}
\examples{
require(scales)
show_col(covalence_palette(palette = "discrete_main"))
}
