% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-covalence.R
\name{generate_pal}
\alias{generate_pal}
\title{Function factory to generate palette}
\usage{
generate_pal(palette, n, type = "discrete", reverse = FALSE)
}
\arguments{
\item{palette}{Palette to use. One of 'complete', 'main', or 'accent'.}

\item{n}{Number of colors wanted from palette.}

\item{type}{Type of palette. Either 'discrete' or 'continuous'.}

\item{reverse}{Should the palette be reversed? Either 'TRUE' or 'FALSE'.}
}
\value{
A palette-generating function.
}
\description{
This function generates a second function, for use in the 'ggplot' scales.
}
\details{
This function is slightly modified from a 2022 \href{https://www.jumpingrivers.com/blog/custom-colour-palettes-for-ggplot2/}{blog post} by Nicola Rennie.
}
