% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-psa.R
\name{plot_psa_scatter}
\alias{plot_psa_scatter}
\title{Plot incremental outcomes on the cost-effectiveness plane}
\usage{
plot_psa_scatter(
  data,
  delta_qalys = NULL,
  delta_costs = NULL,
  currency = "$",
  point_alpha = 0.4,
  point_color = "#249bc9",
  point_shape = 16,
  point_size = 2,
  point_jitter_height = 0,
  point_jitter_width = 0,
  show_wtp = TRUE,
  wtp_value = 1,
  wtp_alpha = 1,
  wtp_color = "#154754",
  wtp_linetype = "dashed",
  wtp_linewidth = 0.5,
  show_mean = TRUE,
  mean_alpha = 1,
  mean_color = "#7a0d66",
  mean_shape = 18,
  mean_size = 3
)
}
\arguments{
\item{data}{A data.frame.}

\item{delta_qalys}{Name of the column that holds incremental QALYs.}

\item{delta_costs}{Name of the column that holds incremental costs.}

\item{currency}{String for the currency symbol. Default is \code{$}.}

\item{point_alpha}{Numeric value for point opacity. Default is \code{0.4}.}

\item{point_color}{Hex value for point color. Default is \verb{#249bc9}.}

\item{point_shape}{Integer value for point shape. Default is \code{16}. See \code{?pch} for available shapes.}

\item{point_size}{Numeric value for point size. Default is \code{2}.}

\item{point_jitter_height, point_jitter_width}{Numeric value for vertical and horizontal jittering. Default is \code{0} (no jittering). Be careful that some key-outcome pairs may not be plotted if jitter too much.}

\item{show_wtp}{Boolean indicating if the willingness-to-pay (WTP) threshold should be displayed (as a line). Default is \code{TRUE}.}

\item{wtp_value}{Numeric value for WTP threshold. Default is \code{1}, a deliberately unrealistic value to remind you to set the value to your specific example.}

\item{wtp_alpha}{Numeric value for WTP threshold line opacity. Default is \code{1}.}

\item{wtp_color}{Hex value for WTP threshold line color. Default is \verb{#154754}.}

\item{wtp_linetype}{A valid specification of the WTP threshold line type. Default is \code{dashed} (\code{2}). See the \href{https://ggplot2.tidyverse.org/reference/aes_linetype_size_shape.html}{ggplot2} documentation for valid options.}

\item{wtp_linewidth}{Numeric value for the WTP threshold line width. Default is \code{0.5}.}

\item{show_mean}{Boolean indicating if the mean of incremental QALYs and costs should be displayed (as a point). Default is \code{TRUE}.}

\item{mean_alpha}{Numeric value for mean point opacity. Default is \code{1}.}

\item{mean_color}{Hex value for mean point color. Default is \verb{#7a0d66}.}

\item{mean_shape}{Integer value for mean point shape. Default is \code{18}. See \code{?pch} for available shapes.}

\item{mean_size}{Numeric value for point size. Default is \code{3}.}
}
\value{
A ggplot object
}
\description{
\code{plot_psa_scatter} takes a data.frame of incremental quality-adjusted life-years (QALYs) and incremental costs, as might be obtained from a probabilistic sensitivity analysis, and plots these data on the incremental cost-effectiveness plane.
}
\details{
Fore more information on the display of incremental outcomes on the cost-effectiveness plane, see, for example, \href{https://pubmed.ncbi.nlm.nih.gov/16623946/}{Fenwick \emph{et al.}, 2006}.
}
\examples{
library(ggplot2)
df <- data.frame(d_qalys = runif(1000, -1, 2),
                 d_costs = runif(1000, -1000, 10000))

plot_psa_scatter(df, d_qalys, d_costs, wtp_value = 20000)
}
