% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-covalence.R
\name{theme_covalence}
\alias{theme_covalence}
\title{The default Covalence \link{ggplot2} theme}
\usage{
theme_covalence(
  base_size = 12,
  base_family = "sans",
  strip_bg_color = "#326aa0",
  strip_text_color = "#FFFFFF"
)
}
\arguments{
\item{base_size}{Base font size, in points. Default is 12.}

\item{base_family}{Base font family. Default is "sans".}

\item{strip_bg_color}{Color to fill strip backgrounds. Default is #326aa0.}

\item{strip_text_color}{Color for strip labels. Default is #FFFFFF (white).}
}
\value{
A ggplot2 \link[ggplot2:theme]{theme} object.
}
\description{
This theme provides the defaults required to generate plots, using \link{ggplot2},
that are aligned with the Covalence Research house style. The theme can be
modified further by using \code{theme()}. Also consider using the various color
palettes to ensure the house style is met.
}
\examples{
library(ggplot2)

ggplot(data = msleep, aes(x = brainwt, y = bodywt)) +
   geom_point() +
   theme_covalence()

ggplot(data = diamonds, aes(x = price, y = carat)) +
   geom_point(color = covalence_colors("lightblue"), alpha = 0.4) +
   facet_wrap(vars(cut)) +
   theme_covalence()
}
